+++ fastdub/audio.py
@@ -3,17 +3,15 @@ from __future__ import annotations
 import math
 import os.path
 from copy import copy
-from math import inf
 from tempfile import TemporaryDirectory
 
 import pydub.silence
-from tqdm import tqdm
 
 from fastdub.ffmpeg_wrapper import FFMpegWrapper
 
 __all__ = ('AudioSegment',
            'speed_change', 'calc_speed_change_ffmpeg_arg',
-           'fit', 'side_chain')
+           'fit')
 
 
 class AudioSegment(pydub.AudioSegment):
@@ -34,12 +32,11 @@ class AudioSegment(pydub.AudioSegment):
 
 def speed_change(audio: AudioSegment, speed_changes: float, allow_copy: bool = True, log_level: str = 'error'
                  ) -> AudioSegment:
-    if speed_changes == 1.:
-        return audio if allow_copy else copy(audio)
-    if speed_changes <= 0.:
+    if speed_changes <= 0:
         raise ValueError(f"Speed cannot be negative ({speed_changes}).\n"
                          "This is usually due to errors in subtitle timecodes.")
-
+    if speed_changes == 1:
+        return audio if allow_copy else copy(audio)
     with TemporaryDirectory() as tmp:
         inp = os.path.join(tmp, 'inp.wav')
         audio.export(inp)
@@ -52,7 +49,8 @@ def speed_change(audio: AudioSegment, speed_changes: float, allow_copy: bool = T
 
 def calc_speed_change_ffmpeg_arg(speed_changes: float) -> str:
     """
-    The given function takes in a float value representing the speed change, and returns a corresponding FFmpeg argument string for changing the speed of an audio or video file.
+    The given function takes in a float value representing the speed change
+    and returns a corresponding FFMpeg argument string for changing the speed of an audio or video file.
     """
     if .5 <= speed_changes <= 100.:
         return f'atempo={speed_changes}'
@@ -73,14 +71,3 @@ def fit(audio: AudioSegment,
         (free - audio_duration) / align
         if audio_duration > need_duration + right_border
         else left_border) + audio
-
-
-def side_chain(sound1: AudioSegment, sound2: AudioSegment,
-               min_silence_len: int = 100, silence_thresh: float = -inf,
-               gain_during_overlay: int = -10
-               ) -> AudioSegment:
-    for start, end in tqdm(
-            pydub.silence.detect_nonsilent(sound2, min_silence_len=min_silence_len, silence_thresh=silence_thresh),
-            desc="Ducking"):
-        sound1 = sound1.overlay(sound2[start:end], start, times=1, gain_during_overlay=gain_during_overlay)
-    return sound1