+++ fastdub/dubber.py
@@ -1,5 +1,6 @@
 from __future__ import annotations
 
+import logging
 import os.path
 import shutil
 from pathlib import Path
@@ -8,11 +9,13 @@ from typing import Sequence
 import rich.align
 from tqdm import tqdm
 
-from fastdub import audio, subtitles, voicer
-from fastdub.audio import calc_speed_change_ffmpeg_arg
+from fastdub import audio, subtitles, voicer, GlobalSettings
+from fastdub.audio import AudioSegment, calc_speed_change_ffmpeg_arg
 from fastdub.ffmpeg_wrapper import FFMpegWrapper
 
 __all__ = ('Dubber', 'VOICER')
+
+from fastdub.subtitles import Line, TimeLabel
 
 VOICER = voicer.Voicer()
 
@@ -22,11 +25,11 @@ class Dubber:
         'fit_align', 'language', 'audio_format',
         'cleanup_audio', 'export_video',
         'ducking',
-        'min_silence_len', 'silence_thresh', 'gain_during_overlay'
+        'sidechain_level_sc', 'sidechain_ffmpeg_params'
     )
 
     def __init__(self, voice: str, language: str, audio_format: str,
-                 ducking: bool, min_silence_len: int, silence_thresh: float, gain_during_overlay: int,
+                 ducking: bool, sidechain_level_sc: float, sidechain_ffmpeg_params: str,
                  fit_align: float = 2., cleanup_audio: bool = True, export_video: bool = True):
         self.language = language
         self.audio_format = audio_format
@@ -35,9 +38,8 @@ class Dubber:
             VOICER.set_voice(voice)
 
         self.ducking = ducking
-        self.min_silence_len = min_silence_len
-        self.silence_thresh = silence_thresh
-        self.gain_during_overlay = gain_during_overlay
+        self.sidechain_level_sc = sidechain_level_sc
+        self.sidechain_ffmpeg_params = sidechain_ffmpeg_params
 
         self.cleanup_audio = cleanup_audio
         self.export_video = export_video
@@ -72,72 +74,68 @@ class Dubber:
         if export_video is None:
             export_video = self.export_video
 
-        rich.print(rich.align.Align(fn, 'center'))
+        logging.info(f'start voice file {fn!r}')
         result_dir = Path(target_sub).parent / '_result'
         result_dir.mkdir(exist_ok=True)
         out_audio_base = result_dir / f'{fn}_{self.language}.{self.audio_format}'
 
         subs = subtitles.parse(target_sub)
-
         default_right_border = 0
         if target_vid and subs:
-            default_right_border = FFMpegWrapper.get_video_duration_ms(target_vid) - subs[-1].ms.end
-        default_right_border = max(default_right_border, 0)
+            default_right_border = FFMpegWrapper.get_video_duration_ms(target_vid)
+        default_right_border = max(default_right_border, end := subs[-1].ms.end)
+        subs += Line(TimeLabel(default_right_border, end, end - default_right_border)),
 
-        progress_total = len(subs)
+        progress_total = len(subs) - 1
 
-        total_duration_ms = 0
         fit_align = self.fit_align
         audio_format = self.audio_format
 
         working_dir = result_dir / '_working_dir'
         working_dir.mkdir(exist_ok=True)
 
-        _audio_filename_format = str(working_dir / ('{0:0>%i}.%s' % (len(str(progress_total)), audio_format))
-                                     ).format
-        filenames: tuple[str] = *(_audio_filename_format(pos) for pos in range(len(subs))),
-        for pos, line in tqdm(enumerate(subs[:-1]),
-                              desc='Creating audio',
-                              total=progress_total - 1, unit='line',
-                              dynamic_ncols=True):
-            ms = subs[pos].ms
-            tts = VOICER.voice(line.text)
+        filenames: tuple[str] = *map(
+            str(working_dir / ('{0:0>%i}.%s' % (len(str(progress_total)), audio_format))).format, range(len(subs))),
+
+        filename_sub = *zip(filenames_striped := filenames[:-1], subs[:-1]),
+        cached_tts = [VOICER.voice(line.text) for fn, line in tqdm(filename_sub,
+                                                                   desc='TTS',
+                                                                   total=progress_total, unit='line',
+                                                                   dynamic_ncols=True)]
+
+        total_duration_ms = 0
+        for pos, ((tts_fn, line), cached) in tqdm(
+                enumerate(zip(filename_sub, cached_tts), 1),
+                desc='Fitting',
+                total=progress_total, unit='line',
+                **GlobalSettings.tqdm_kwargs):
+            ms = line.ms
             new_audio = audio.fit(
-                tts,
+                AudioSegment.from_file(cached, 'wav'),
                 ms.start - total_duration_ms,
                 ms.duration,
-                subs[pos + 1].ms.start - ms.end,
+                subs[pos].ms.start - ms.end,
                 fit_align
             )
             total_duration_ms += new_audio.duration_ms
-            new_audio.export(filenames[pos], audio_format)
-        ms = subs[-1].ms
-        tts = VOICER.voice(subs[-1].text)
-        new_audio = audio.fit(tts,
-                              ms.start - total_duration_ms,
-                              ms.duration,
-                              default_right_border,
-                              fit_align)
-        total_duration_ms += new_audio.duration_ms
-        new_audio.export(filenames[-1], audio_format)
-
-        with open(list_file := os.path.join(working_dir, 'list.txt'), 'w') as f:
-            f.write('\n'.join(f"file '{fn}'" for fn in filenames))
-        max_duration = ms.end
+            new_audio.export(tts_fn, audio_format)
+        with open(list_file := working_dir / 'list.txt', 'w') as f:
+            f.writelines(f"file '{fn}'\n" for fn in filenames_striped)
+        max_duration = subs[-1].ms.end
 
         ffmpeg_concat_args = ()
         if total_duration_ms > max_duration:
             change_speed = total_duration_ms / max_duration
-            rich.print(f'Changing audio speed to {change_speed:g}')
+            logging.info(f'changing audio speed {change_speed:g}')
             ffmpeg_concat_args += '-af', calc_speed_change_ffmpeg_arg(change_speed)
         else:
             ffmpeg_concat_args += '-c', 'copy'
 
-        fitted_audio_file = str(out_audio_base.with_stem(f'_{out_audio_base.stem}'))
-        rich.print('Concatenating parts...', flush=True)
+        temp_audio_file = str(out_audio_base.with_stem(f'_{out_audio_base.stem}'))
+        logging.info('concatenating parts...')
         FFMpegWrapper.convert('-f', 'concat', '-safe', '0', '-i', list_file, *ffmpeg_concat_args,
-                              fitted_audio_file)
-        cur_audio = audio.AudioSegment.from_file(fitted_audio_file)
+                              temp_audio_file)
+        cur_audio = audio.AudioSegment.from_file(temp_audio_file)
         shutil.rmtree(working_dir, ignore_errors=True)
 
         if total_duration_ms != max_duration:
@@ -147,20 +145,23 @@ class Dubber:
 
         result_out_audio = str(out_audio_base)
         if target_vid:
+            cur_audio.export(temp_audio_file, format='wav')
             if self.ducking:
-                audio.side_chain(audio.AudioSegment.from_file(target_vid),
-                                 cur_audio,
-                                 self.min_silence_len, self.silence_thresh, self.gain_during_overlay
-                                 ).export(result_out_audio)
+                logging.info('sidechain')
+                FFMpegWrapper.sidechain(target_vid,
+                                        temp_audio_file,
+                                        result_out_audio,
+                                        self.sidechain_level_sc,
+                                        self.sidechain_ffmpeg_params)
             else:
-                audio.AudioSegment.from_file(target_vid).overlay(cur_audio,
-                                                                 gain_during_overlay=self.gain_during_overlay
-                                                                 ).export(result_out_audio)
+                logging.info('amix')
+                FFMpegWrapper.amix(target_vid, temp_audio_file, out=result_out_audio)
+                # audio.AudioSegment.from_file(target_vid).overlay(cur_audio).export(result_out_audio)
             if export_video:
                 FFMpegWrapper.save_result_data(target_vid, result_out_audio, target_sub,
-                                               result_dir / f'{fn}_{self.language}.mp4')
+                                               result_dir / f'{fn}_{self.language}.mkv')
                 if cleanup_audio:
                     os.remove(result_out_audio)
         else:
             cur_audio.export(result_out_audio)
-        os.remove(fitted_audio_file)
+        os.remove(temp_audio_file)